requirements.txt
fastapi
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
python-multipart
aiofiles
fastapi-utils